+++ samples/balloon/balloon.py
@@ -42,9 +42,8 @@ if ROOT_DIR.endswith("samples/balloon"):
 
 # Import Mask RCNN
 sys.path.append(ROOT_DIR)
-from config import Config
-import utils
-import model as modellib
+from mrcnn.config import Config
+from mrcnn import model as modellib, utils
 
 # Path to trained weights file
 COCO_WEIGHTS_PATH = os.path.join(ROOT_DIR, "mask_rcnn_coco.h5")
